disambiguated exiting to menu, made bad uint joke
it's important that the reader understands that the script still works on resets. I don't want people thinking they have to close and re-open everything just to do a new run.

it's also very important that I make a deadpan joke about the fact that counter can't be a 16-bit integer, because the script would crash during the third run.